extract JIT dllmain.cpp and eliminate clrjit_mergeable
