+++ src/coreclr/src/jit/ee_il_dll.cpp
@@ -37,9 +37,6 @@ FILE* jitstdout = nullptr;
 ICorJitHost*   g_jitHost        = nullptr;
 static CILJit* ILJitter         = nullptr; // The one and only JITTER I return
 bool           g_jitInitialized = false;
-#ifndef FEATURE_MERGE_JIT_AND_ENGINE
-HINSTANCE g_hInst = nullptr;
-#endif // FEATURE_MERGE_JIT_AND_ENGINE
 
 /*****************************************************************************/
 
@@ -153,33 +150,6 @@ void jitShutdown(bool processIsTerminating)
 
     g_jitInitialized = false;
 }
-
-#ifndef FEATURE_MERGE_JIT_AND_ENGINE
-
-extern "C" DLLEXPORT BOOL WINAPI DllMain(HANDLE hInstance, DWORD dwReason, LPVOID pvReserved)
-{
-    if (dwReason == DLL_PROCESS_ATTACH)
-    {
-        g_hInst = (HINSTANCE)hInstance;
-        DisableThreadLibraryCalls((HINSTANCE)hInstance);
-    }
-    else if (dwReason == DLL_PROCESS_DETACH)
-    {
-        // From MSDN: If fdwReason is DLL_PROCESS_DETACH, lpvReserved is NULL if FreeLibrary has
-        // been called or the DLL load failed and non-NULL if the process is terminating.
-        bool processIsTerminating = (pvReserved != nullptr);
-        jitShutdown(processIsTerminating);
-    }
-
-    return TRUE;
-}
-
-HINSTANCE GetModuleInst()
-{
-    return (g_hInst);
-}
-
-#endif // !FEATURE_MERGE_JIT_AND_ENGINE
 
 /*****************************************************************************/
